rag-orchestrator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# OpenAI
openai==1.54.0
httpx==0.27.0

# YAML
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
